+++ NEWS.md
@@ -13,6 +13,7 @@
 * Removed `get_adm_boundaries` given depreciation of `raster`
 * Updated `geo_neighbors` to require source as an external dataset
 * Adding `get_nepolygons` as a common way of extracting naturalearth database
+* Updated `get_hexbins` to require spdf as input + redius
 
 # gisr 0.3.0